Hijack sbt shell, remember last command
That's much more robust than inspecting the state, because that lets us
re-run exactly what the user typed in the previous run. For instance,
the previous implementation didn't work if the user typed

    > ;cmd1 ;cmd2

Because the two commands would be separated. It also didn't work with
commands that inject more commands into the state.

Those cases should be handled correctly now.